Split SSO provider registry into one file per provider
Move lib/onetime/sso_provider_registry.rb to
lib/onetime/sso_provider/registry.rb and extract each provider
definition into its own file (oidc.rb, entra.rb, google.rb, github.rb)
under lib/onetime/sso_provider/, each exposing a frozen DEFINITION
hash. The registry requires them and fixes their order, keeping
DEFINITIONS and .fetch unchanged for consumers.

Rename the module Onetime::SsoProviderRegistry to
Onetime::SsoProvider::Registry and update both consumers (AuthConfig,
the auth app omniauth feature), the specs, docs, and .env.reference.
Definition data is unchanged — verified old and new DEFINITIONS are
equal field-for-field, including env-built strategy options.

Prepares for additional OmniAuth strategies: adding a provider is now
a self-contained file plus a one-line registry entry.
