+++ .env.reference
@@ -764,7 +764,7 @@ GOOGLE_CLIENT_SECRET=
 # Login-button ordering for SSO providers: a comma/space-separated list of
 # route names (e.g. "google, github, oidc"). Listed providers appear first
 # in the given order; unlisted ones keep their registry order after them.
-# Unset = registry order (lib/onetime/sso_provider_registry.rb).
+# Unset = registry order (lib/onetime/sso_provider/registry.rb).
 #SSO_PROVIDER_ORDER=
 
 # Multi-tenant SSO fallback policy (auth:
